+++ client-device-connection-infinispan/src/main/java/org/eclipse/hono/deviceconnection/infinispan/client/HotrodBasedDeviceConnectionInfo.java
@@ -121,6 +121,11 @@ public final class HotrodBasedDeviceConnectionInfo implements DeviceConnectionIn
         Objects.requireNonNull(deviceId);
 
         return cache.get(getGatewayEntryKey(tenantId, deviceId))
+                .recover(t -> {
+                    LOG.debug("failed to find last known gateway for device [tenant: {}, device-id: {}]",
+                            tenantId, deviceId, t);
+                    return Future.failedFuture(new ServerErrorException(HttpURLConnection.HTTP_INTERNAL_ERROR, t));
+                })
                 .compose(gatewayId -> {
                     if (gatewayId == null) {
                         LOG.debug("could not find last known gateway for device [tenant: {}, device-id: {}]", tenantId,
@@ -131,11 +136,6 @@ public final class HotrodBasedDeviceConnectionInfo implements DeviceConnectionIn
                                 deviceId, gatewayId);
                         return Future.succeededFuture(getLastKnownGatewayResultJson(gatewayId));
                     }
-                })
-                .recover(t -> {
-                    LOG.debug("failed to find last known gateway for device [tenant: {}, device-id: {}]",
-                            tenantId, deviceId, t);
-                    return Future.failedFuture(new ServerErrorException(HttpURLConnection.HTTP_INTERNAL_ERROR, t));
                 });
     }